old-tests/submission/Microsoft/video/video_020.htm

<!DOCTYPE HTML>
<html>
<head>
    <title>HTML5 Media Elements: 'pause' method sets 'paused' attribute to true.</title>
    <meta content="text/html; charset=UTF-8" http-equiv="Content-Type">
    <link rel="author" title="Microsoft" href="http://www.microsoft.com/" />
    <link rel="help" href="http://www.w3.org/TR/html5/video.html#dom-media-pause" />
    <meta name="assert" content="'pause' method sets 'paused' attribute to true." />
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script type="text/javascript" src="../common/media.js"></script>
    <script type="text/javascript">
        var pause = async_test("'pause' method sets 'paused' attribute to true")
        var play = async_test("'paused' attribute is initially set to false")
        
        function do_playing(event)
        {
            play.step(function(){ assert_false(document.getElementById("video0").paused)});
            play.done();
            document.getElementById("video0").addEventListener("pause", do_pause, false);
            document.getElementById("video0").pause();
        }

        function do_pause(event)
        {
            pause.step(function(){ assert_true(document.getElementById("video0").paused)});
            pause.done();
            clearTimeout(timer);
        }
        
        function timeout()
        {
            play.step(function(){ assert_false(true, "'onplaying' event never fired")});
            pause.step(function(){ assert_false(true, "'pause' event never fired")});
            play.done();
            pause.done();
        }
  
        var timer = setTimeout(timeout, 10000);
        
    </script>
</head>
<body>
  <div id=log></div>
  
  <script type="text/javascript">
  document.write(
      "<video style='display : none' id='video0' src='" + getVideoURI("/media/movie_5") + "'" +
      " onplaying='do_playing(event);'" +
      " autoplay>"
      );
  
  </script>
  Your browser does not support media elements.
  </video>
</body>
</html>
